finsihed work with mobile view for kanban
