+++ app/assets/main/react_components/User/MobileKanbanContainer.jsx
@@ -1,9 +1,22 @@
-import React from "react";
+import React, { useState } from "react";
 import KanbanActionContainer from "./KanbanActionContainer.jsx";
-import { useMediaQuery } from "react-responsive";
 
-const MobileKanbanContainer = ({ categories }) => {
-  const isTabletOrMobile = useMediaQuery({ query: "(max-width: 1224px)" });
+const MobileKanbanContainer = ({ categories, climateActions }) => {
+  const getTotalAmountOfAcceptedActionsForUser = () => {
+    let amountActionsAccepted = 0;
+    climateActions.map((action) => {
+      if (action.accepted === true) {
+        amountActionsAccepted++;
+      }
+    });
+    return amountActionsAccepted;
+  };
+
+  const [showMobileKanban, setShowMobileKanban] = useState(false);
+
+  const setChecked = () => {
+    setShowMobileKanban(!showMobileKanban);
+  };
 
   return (
     <>
@@ -11,16 +24,22 @@ const MobileKanbanContainer = ({ categories }) => {
       <div className="fixed top-20 z-30 bg-white w-full overflow-hidden h-0 toggler-checked:h-screen transition-size duration-500">
         <KanbanActionContainer categories={categories} />
       </div>
-
-      {/**Prevent background site from scrolling */}
-      {isTabletOrMobile && (
-        <label
-          htmlFor="mobileKanban"
-          className="fixed top-0 right-0 mr-20 mt-6 lg:mr-48 lg:right-10 t:mr-17 t:right-3 z-50"
-        >
-          <i className={`fas fa-2x ${"fa-globe-americas"}`}></i>
-        </label>
-      )}
+      <label
+        htmlFor="mobileKanban"
+        className="fixed top-0 right-0 mr-20 mt-6 lg:mr-48 lg:right-10 t:mr-17 t:right-3 z-50"
+      >
+        <i
+          className={`fas fa-2x ${
+            showMobileKanban ? "fa-globe-europe" : "fa-globe-americas"
+          }`}
+          onClick={setChecked}
+        ></i>
+        <div className="fas rounded-full h-5 w-5 bg-green-tint-2 border border-gray-accent -mt-1 -ml-3 absolute focus:outline-none">
+          <div className="mb-2 text-white text-center">
+            {getTotalAmountOfAcceptedActionsForUser()}
+          </div>
+        </div>
+      </label>
     </>
   );
 };
